+++ crates/pathfinder/src/sync/events.rs
@@ -27,31 +27,7 @@ use crate::sync::stream::ProcessStage;
 
 /// Returns the first block number whose events are missing in storage, counting
 /// from genesis
-pub(super) async fn next_missing(
-    storage: Storage,
-    head: BlockNumber,
-) -> anyhow::Result<Option<BlockNumber>> {
-    spawn_blocking(move || {
-        let mut db = storage
-            .connection()
-            .context("Creating database connection")?;
-        let db = db.transaction().context("Creating database transaction")?;
-
-        if let Some(highest) = db
-            .highest_block_with_all_events_downloaded()
-            .context("Querying highest block with events")?
-        {
-            Ok((highest < head).then_some(highest + 1))
-        } else {
-            Ok(Some(BlockNumber::GENESIS))
-        }
-    })
-    .await
-    .context("Joining blocking task")?
-}
-
-#[cfg(feature = "aggregate_bloom")]
-pub(super) fn next_missing_aggregate(
+pub(super) fn next_missing(
     storage: Storage,
     head: BlockNumber,
 ) -> anyhow::Result<Option<BlockNumber>> {